+++ lib/ghamma/cli/duration.rb
@@ -1,6 +1,7 @@
 require "csv"
 require "dry/cli"
 require "http"
+require "tty/progressbar"
 
 require_relative "../github_api_client"
 
@@ -12,14 +13,23 @@ module Ghamma
       argument :owner, required: true, desc: "The user or organisation to whom the repo belongs"
       argument :repo, required: true, desc: "The repo to whom the workflow belongs"
       argument :workflow, required: true, desc: "The filename for the desired workflow, e.g. tests.yml"
-      option :since, default: "1970-01-01", desc: "Fetch workflow runs since this date"
+      option :since, desc: "Fetch workflow runs since this date"
       option :output, desc: "Optional file to which to output results, defaults to STDOUT"
 
-      def call(owner:, repo:, workflow:, since:, output: nil)
-        durations = GithubApiClient.new(owner: owner, repo: repo, token: ENV["GH_TOKEN"])
-          .fetch_workflow_duration_history(workflow, since)
+      def call(owner:, repo:, workflow:, since: nil, output: nil)
+        workflow_statement = "Fetching the durations of all workflow runs of #{workflow}"
+        since_statment = "since #{since}" if since
 
-        puts "Workflows fetchced, generating output"
+        puts [workflow_statement, since_statment].join(" ")
+
+        progressbar = TTY::ProgressBar.new("Fetching workflow runs [:bar] :current/:total ETA: :eta", total: nil)
+
+        durations = GithubApiClient.new(owner: owner, repo: repo, token: ENV["GH_TOKEN"])
+          .fetch_workflow_duration_history(workflow, since, progressbar)
+
+        progressbar.finish
+
+        puts "Generating output"
 
         durations_output = CSV.generate do |csv|
           csv << ["ID", "Date", "Duration"]
@@ -31,6 +41,7 @@ module Ghamma
         if output
           File.write(output, durations_output)
         else
+          puts
           puts durations_output
           puts
         end